refactor(map): remove dead mapController.ts, wire controls to optimized controller
- Add zoomAllFeatures() to optimizedMapController so ZoomAllControl works
- Update ZoomAllControl and LocationControl imports to optimizedMapController
- Delete mapController.ts (was unreachable; SightingsMapView already used
  optimizedMapController exclusively)
